refactor: extract util methods into pure functions
Preparation for writing some tests
