+++ src/index.ts
@@ -2,6 +2,7 @@ import { SSM } from 'aws-sdk';
 import chalk from 'chalk';
 
 import { ServerlessInstance, ServerlessOptions, SSMParam } from './types';
+import { compareParams, evaluateEnabled, validateParams } from './util';
 
 const unsupportedRegionPrefixes = [
   'ap-east-1',    // Hong Kong - region disabled by default
@@ -94,7 +95,7 @@ class ServerlessSSMPublish {
    */
   private initializeVariables(): void {
     if (!this.initialized) {
-      this.enabled = this.evaluateEnabled();
+      this.enabled = evaluateEnabled(this.serverless.service.custom, this.throwError.bind(this)); // tslint:disable-line:no-unsafe-any
 
       if (this.enabled) {
         const credentials = this.provider.getCredentials(); // tslint:disable-line:no-unsafe-any
@@ -106,7 +107,7 @@ class ServerlessSSMPublish {
         const credentialsWithRegion = { ...credentials, region: this.region };
         this.ssm = new this.provider.sdk.SSM(credentialsWithRegion); // tslint:disable-line:no-unsafe-any
 
-        this.params = this.validateParams();
+        this.params = validateParams(this.serverless.service.custom.ssmPublish, this.throwError.bind(this), this.log.bind(this)); // tslint:disable-line:no-unsafe-any
 
         unsupportedRegionPrefixes.forEach((unsupportedRegionPrefix) => {
           if (this.region.startsWith(unsupportedRegionPrefix)) {
@@ -121,77 +122,6 @@ class ServerlessSSMPublish {
   }
 
   /**
-   * Determines whether this plugin should be enabled.
-   *
-   * This method reads the ssmPublish property "enabled" to see if this plugin should be enabled.
-   * If the property's value is undefined, a default value of true is assumed (for backwards compatibility).
-   * If the property's value is provided, this should be boolean, otherwise an exception is thrown.
-   */
-  private evaluateEnabled(): boolean {
-    if (!this.serverless.service.custom?.ssmPublish) {
-      this.throwError('Plugin configuration is missing.');
-    }
-
-    const enabled = this.serverless.service.custom.ssmPublish.enabled;
-    // const err = new Error(`Ambiguous value for "enabled": '${enabled}'`);
-
-    // Enabled by default
-    if (enabled === undefined) {
-      return true;
-    }
-
-    switch (typeof enabled) {
-      case 'boolean':
-        return enabled;
-      case 'string':
-        if (enabled === 'true') return true;
-        if (enabled === 'false') {
-          return false;
-        }
-        this.throwError(`Ambiguous value for "enabled": '${enabled}'`);
-        return false;
-      default:
-        this.throwError(`Ambiguous value for "enabled": '${enabled}'`);
-        return false;
-    }
-  }
-
-  /**
-   * Validates params passed in serverless.yaml
-   * Throws error if no params or incorrect syntax.
-   * Might want to call this in evaluateEnabled?
-   * Need to add some more validation here
-   */
-  private validateParams(): SSMParam[] | undefined {
-    if (!this.serverless.service?.custom?.ssmPublish?.params || !this.serverless.service?.custom?.ssmPublish?.params.length) {
-      this.throwError('No params defined');
-    }
-
-    const validateParam = (param: SSMParam) => {
-      const maxNameLength = 1011; // needs to account for ARN stuff being added
-      const maxDescriptionLength = 1024;
-      const maxDepth = 15;
-      if (!['path', 'value'].every((requiredKey: string) => Object.keys(param).includes(requiredKey)))
-        this.throwError('Path and Value are required fields for params');
-      if (typeof param.secure !== 'boolean') {
-        this.log(chalk.redBright(`Param at path ${param.path} should pass Secure as boolean value`));
-      }
-      if (
-          param.path.match(/^(aws|ssm)/gi) || // check if name begins with illegal patterns (aws/ssm)
-          param.path.match(/[^a-zA-Z0-9_.\-\/]/g) || // check if name contains illegal characters
-          (param.path.match(/\//g) || []).length > maxDepth ||
-          param.path.length > maxNameLength
-        )
-        this.throwError(`Param ${param.path} name doesn't match AWS constraints`);
-      if (param.description && param.description.length > maxDescriptionLength)
-        this.throwError(`Param ${param.path} description is too long`);
-      return { ...param, Secure: !!param.secure };
-    };
-
-    return this.serverless.service.custom.ssmPublish.params?.map(validateParam);
-  }
-
-  /**
    * Checks whether parameters exist in SSM and if they've been changed
    * Stores arrays of changed/unchanged/new Parameters on class
    */
@@ -199,20 +129,7 @@ class ServerlessSSMPublish {
     if (!this.params) return;
     const retrievedParameters = await this.ssm.getParameters({ Names: this.params.map((param) => param.path), WithDecryption: true}).promise();
 
-    const { nonExistingParams, existingChangedParams, existingUnchangedParams } = this.params.reduce< { nonExistingParams: SSMParam[]; existingChangedParams: SSMParam[]; existingUnchangedParams: SSMParam[] }>((acc, curr) => {
-      const existingParam = retrievedParameters.Parameters?.find((param) => param.Name === curr.path);
-      if (!existingParam) {
-        acc.nonExistingParams.push(curr);
-        return acc;
-      }
-      if (existingParam.Value === curr.value) {
-        acc.existingUnchangedParams.push(curr);
-        return acc;
-      }
-      acc.existingChangedParams.push(curr);
-      return acc;
-    }
-    , { nonExistingParams: [], existingChangedParams: [], existingUnchangedParams: [] });
+    const { nonExistingParams, existingChangedParams, existingUnchangedParams } = compareParams(this.params, retrievedParameters);
 
     this.logIfDebug(`New param paths:\n\t${nonExistingParams.map((param) => param.path).join('\n\t')}`);
     this.logIfDebug(`Changed param paths:\n\t${existingChangedParams.map((param) => param.path).join('\n\t')}`);
